Moved nav bar into correct folder
